move SelectWallet out of SignUp flow
