+++ src/routes/+page.ts
@@ -1,0 +1,8 @@
+import {getRpcEndpointFromURL} from "$lib/util";
+
+export function load({ url }: any) {
+  return {
+    ...getRpcEndpointFromURL(url),
+    schemas: url.searchParams.get('schemas'),
+  };
+}
