+++ src/routes/signin/+page.svelte
@@ -1,11 +1,2 @@
 <script lang="ts">
-  import SelectWallet from '$components/SelectWallet.svelte';
 </script>
-
-<div class="text-center mt-7">
-  <h1 class="mt-4 text-2xl text-aqua text-center">Sign in</h1>
-  <p>Choose a wallet to connect and sign in.</p>
-</div>
-<div class="mt-8">
-  <SelectWallet />
-</div>
